feat(raster-core): full frame API, instancing, x-ray deep pass; fix morph rest weight
Root-cause of the empty mesh pipeline under wgpu 30/Metal: WebGPU NDC z is
[0, 1] (Vulkan-style), not OpenGL [-1, 1] — the test projection clipped
every primitive. While bisecting, found and fixed a real renderer bug: the
morph rest weight was 0, so mix(zero_dummy, pos, 0) collapsed every
non-morphing mesh to the origin (bench timings masked it). Rest weight is
now 1 everywhere (init, retire, instances), making the slot-1 dummy inert.

raster-core:
- Frame API composing the full pipeline set in browser order: grid,
  opaque meshes, instanced meshes, transparent meshes, deep-selection
  x-ray overlay (depth Always, on top), edges, lines, selection overlay.
- Instancing: per-instance Obj records in a storage buffer pool + shared
  geometry (instanced mesh/edge pipelines).
- Edge draws, per-vertex colored line/overlay batches, fullscreen grid,
  write_ppm snapshots; PPM artifacts gitignored.
- tests: 14 green (naga validation, GPU morph, frame composition with
  pixel assertions, determinism), PPM snapshots visually verified.
- examples/bench.rs: 200 meshes ~4.8k tris @512x512, ~1.3 ms/frame
  (~750 fps) on M4 Max/Metal; README documents ownership and contracts.

+++ crates/geometry-bridge/src/lib.rs
@@ -1839,7 +1839,7 @@ impl CadMeshBuffer {
     }
 }
 /// Typed native entry point shared with the WASM import adapter.
-pub fn import_cad_mesh(stride: usize, vertices: &[f32], indices: &[u32]) -> Result<u32> {
+pub fn import_cad_mesh(stride: usize, vertices: &[f64], indices: &[u32]) -> Result<u32> {
     mesh::import_buffers(stride, vertices, indices)
 }
